Julie_bot/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.10.1
bleach==6.1.0
blessed==1.20.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
coloredlogs==15.0.1
coredis==4.16.0
cryptography==41.0.5
dataclasses-json==0.5.14
datasets==2.14.6
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.6
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.2.6
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-environ==0.10.0
django-extensions==3.0.3
django-flags==5.0.12
django-heroku==0.3.1
django-jsonform==2.17.4
django-picklefield==3.1
django-redis==5.2.0
django-rq==2.7.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
docstring-parser==0.15
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.13.1
FLAML==2.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
geojson==2.5.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
googlesearch-python==1.2.3
greenlet==3.0.1
grpcio==1.59.2
gunicorn==20.0.4
h11==0.14.0
hiredis==2.2.3
httptools==0.6.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
inquirer==3.1.3
install==1.3.5
instructor==0.2.9
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kubernetes==28.1.0
langchain==0.0.327
langsmith==0.0.54
litellm==0.8.6
llama-index==0.8.56
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.1
ooba==0.0.21
open-interpreter==0.1.10
openai==0.28.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
posthog==3.0.2
prettytable==3.9.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyautogen==0.1.14
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.16.1
PyJWT==2.8.0
pymemgpt==0.1.15
Pympler==1.0.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyowm==3.3.0
pypdf==3.17.0
PyPika==0.48.9
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-http-client==3.3.7
python-ulid==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
questionary==2.0.1
readchar==4.0.5
redis==4.6.0
redis-om==0.2.1
redis-search-django==0.1.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rq==1.15.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
selenium==4.14.0
sendgrid==6.4.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
tokentrim==0.1.12
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.1
trio==0.22.2
trio-websocket==0.11.1
triton==2.1.0
typer==0.9.0
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.9
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.9
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
wget==3.2
whitenoise==5.1.0
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.2
yaspin==3.0.1
zipp==3.17.0
